add option to sort map keys in convert-stack-to-iidy (default=true)
also factor this code out into a separate module.
